QueueCards: Extract entry & bookmark cards to seperate files
